requirements.txt
aiogram
pandas
sqlalchemy
python-dotenv
prometheus-client
redis
asyncpg